refactor(image-plugin-redesign): PR Feedback changes - Improve UX and code
