+++ packages/editor/src/plugins/image/editor.tsx
@@ -9,7 +9,7 @@ import { cn } from '@serlo/frontend/src/helper/cn'
 import { useEffect, useRef, useState } from 'react'
 
 import type { ImageProps } from '.'
-import { ImageSelectionScreen } from './image-selection-screen'
+import { ImageSelectionScreen } from './components/image-selection-screen'
 import { ImageRenderer } from './renderer'
 import { ImageToolbar } from './toolbar'
 import { TextEditorConfig } from '../text'
